+++ QAccelPlot/src/transitions/MorphTransition.cpp
@@ -16,8 +16,8 @@ MorphTransition::MorphTransition(QObject* parent)
 {
 }
 
-void MorphTransition::interpolate(const float easedProgress, const std::vector<float>& fromData, const int fromPointCount, const std::vector<float>& toData,
-    const int toPointCount, std::vector<float>& outData, int& outPointCount)
+void MorphTransition::interpolate(const double easedProgress, const std::vector<double>& fromData, const int fromPointCount, const std::vector<double>& toData,
+    const int toPointCount, std::vector<double>& outData, int& outPointCount)
 {
     if (toPointCount <= 0) {
         outPointCount = 0;
